% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffx.R
\name{ffx_search}
\alias{ffx_search}
\title{Search Fairfax County open datasets by keyword}
\usage{
ffx_search(query)
}
\arguments{
\item{query}{Character string to search for (case-insensitive)}
}
\value{
tibble (same schema as ffx_datasets)
}
\description{
Search Fairfax County open datasets by keyword
}
